Refactor: Split useUserProgress.ts into smaller files
Refactor useUserProgress.ts into smaller, more manageable files to improve code organization and maintainability. Ensure all functionality remains intact and remove any unused imports or files.
