+++ src/hooks/user-progress/types.ts
@@ -1,0 +1,8 @@
+
+export interface UserCourseProgress {
+  course_id: string;
+  progress_percentage: number;
+  is_completed: boolean;
+  is_saved: boolean;
+  last_listened_at: string;
+}
